% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrel_porci.R
\name{rgl}
\alias{rgl}
\title{Regresión lineal para calcular cantidad de lechones en próximo parto}
\usage{
rgl(pc, npa, ec, ca, cl, datos)
}
\arguments{
\item{datos}{(vector) Datos reales para entrenamiento de regresión lineal.}

\item{p_cerda}{(vector) Pesos de cerda landrace.}

\item{no_partos}{(vector) Número de partos anteriores.}

\item{edad}{(vector) Edad de la cerda en meses}

\item{con_ali}{(vector) Condición alimenticia. 1 = Óptima, 0 = Restringida}

\item{cant_lech}{(vector) Promedio de lechones de cerda landrace.}
}
\value{
Cantidad de lechones en próximo parto.
}
\description{
Regresión lineal para calcular cantidad de lechones en próximo parto
}
